+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## 1.43.0
 
+- Added parameter `files` to `Wizaplace\SDK\Company\CompanyService::registerC2Ccompany`
+
 ### New features
 
 - Added optional parameter `$declineReason` to `\Wizaplace\SDK\Vendor\Order\OrderService::declineOrder`
@@ -27,6 +29,8 @@
 
 - Added `\Wizaplace\SDK\Order\OrganisationOrderService::getOrder`
 - Added attribute `hidden` to `\Wizaplace\SDK\Organisation\OrganisationBasket`
+- Added parameter `files` to `Wizaplace\SDK\Company\CompanyService::registerC2Ccompany`
+- Added `\Wizaplace\SDK\Cms\MenuItem::isTargetBlank`
 
 ## 1.41.0